PLAT-43266: Don't pass random cookies when proxying REST calls.
Some customers, with on-prem Search UI in particular, are using UIs
which are either themselves passing lots of cookies/cookie data or
are passing cookies accumulated 'upstream' (e.g. if the Search UI is
running at searchui.example.com, they're also getting all of the
example.com cookies passed, too.

This fix adds 'Cookie' to the list of headers that are not coppied when
making the request to the 'real' back-end server. (Previously, this
list included only the 'Host' header.)

If trace logging is enabled for the RestProxy class, then the fact that
the Cookie headers are being omitted explicitly is logged (along with
the same informaiton about the Host header).