PLAT-44732: Fix SAML infinite looping
When a user who has an account on the IdP system but doesn't have
permission for the given app tried to log in, they would see an
infinite loop of SAML requests insteaed of a nice error. AFAIK,
this only occurred in ADFS/Azure AD.

This fixes that by upgrading the Spring Boot SAML library to the
latest version.